Controladores lenguage, categoria y actor litos

+++ Catalogo/src/main/java/com/catalogo/IServices/ILanguageService.java
@@ -2,6 +2,6 @@ package com.catalogo.IServices;
 
 import com.catalogo.domain.Language;
 
-public interface ILanguageService extends GenericServiceInterface<Language, Byte>{
+public interface ILanguageService extends GenericServiceInterface<Language, Integer>{
 
 }
